requirements.txt
fastapi==0.111.0
httpx==0.27.0
pydantic==2.7.0
uvicorn[standard]
pytest
pytest-asyncio
redis
fastapi-cache2
